LLM_Files/requirements.txt
ollama
thefuzz
pandas
yfinance
pymongo
python-dotenv
requests
